+++ sdk/containerinstance/azure-mgmt-containerinstance/CHANGELOG.md
@@ -1,5 +1,17 @@
 # Release History
 
+## 3.0.0 (2021-07-20)
+
+**Features**
+
+  - Model ContainerHttpGet has a new parameter http_headers
+  - Model LogAnalytics has a new parameter workspace_resource_id
+  - Added operation ContainersOperations.attach
+
+**Breaking changes**
+
+  - Operation ContainersOperations.list_logs has a new signature
+
 ## 2.0.0 (2020-06-24)
 
 **Features**